Check if runfolder exists before attempting to set state (#29)
* Fix README formatting

* Add test for set_runfolder_state

* Raise exception if runfolder does not exist

Fix #28

+++ runfolder/services.py
@@ -210,12 +210,22 @@ class RunfolderService:
 
     @staticmethod
     def set_runfolder_state(runfolder, state):
-        """Sets the state of a runfolder"""
+        """
+        Sets the state of a runfolder
+
+        :raises DirectoryDoesNotExist
+        """
         validate_state(state)
         arteria_dir = os.path.join(runfolder, ".arteria")
         state_file = os.path.join(arteria_dir, "state")
+
+        if not os.path.exists(runfolder):
+            raise DirectoryDoesNotExist(
+                    "Directory does not exist: '{0}'".format(runfolder))
+
         if not os.path.exists(arteria_dir):
             os.makedirs(arteria_dir)
+
         with open(state_file, 'w') as f:
             f.write(state)
